+++ NEWS.md
@@ -1,5 +1,7 @@
 # ctfs 0.0.0.9005
 
+- Imports package date "depends"
+
 todo
 
 - keep different branches one for enhances and another one for fixes, reports or tests.